Fix double release
The Camera::Device::close() released the capture object, which was
also done in the destructor for Camera::Device as well. Could not
remove this action from the close method becase the user progam
calles this function via a different path and multiple calls
could result in objects not being freed prpperly.

This patch also ensures that the VideoCapture_K can tbe constructed
with either a camera number or a path.