Tello support (#146)
* Remove low resolution restriction

* Putting the fixed version of wait_for_light back

A pull request got merged that removed this because it was from an earlier point in the master branch. 
This is putting the wait_for_light function back that wasn't intended to be removed.

* Camera speed test now evaluates all resolutions

* Set camera defaults to HIGH_RES

Testing has shown that the Wombat can handle 640x480.

Have not tested higher resolutions yet :-)

* Libwallaby - tello support

Initial commit for tello support

Issues:

(1) need to merge set connection to the tello

(2) handle termination conditions

(3) depriacte ffmpeg stuff needs to be updated

(4) examples need to be added

Signed-off-by: Eugene Myers <cedarhouse@comcast.net>

* Set camera defaults to HIGH_RES

Testing has shown that the Wombat can handle 640x480.

Have not tested higher resolutions yet :-)

* Create main.yml

* Update and rename main.yml to cmake_build.yml

* Update cmake_build.yml

The previous action template did not copy over correctly.

* Add with_vision_support compile flag

Enables builds without vision dependencies (opencv, zbar, etc.)

* Create bob.yml

* Delete bob.yml

* Delete cmake_build.yml

* Update README.md

* Sudo necessary

* Emscripten update registers only when changed

* For JS, update multi-byte registers in batches

* Libwallaby - tello support

Signed-off-by: Eugene Myers <cedarhouse@comcast.net>

* make define TELLO as default camera for BOTUI/TELLO testing

This patch should be revoked before merging with master

Signed-off-by: Eugene Myers <cedarhouse@comcast.net>

* Update tellosupport_readme

tellosupport_readme has been updated to:
	(1) Account for function changes suggested in the first review
	(2) Documented an easier method for set TELLO as default
	    camera and faciltating its removal
	(3) Add suggested procedure for testing Tello with Botui.

Signed-off-by: Eugene Myers <cedarhouse@comcast.net>

* Make classes camelcase

Signed-off-by: Eugene Myers <cedarhouse@comcast.net>

* Update camera.cpp

Turn off BOTUI_TELLO_TEST - This only needs to be on when testing Tellos with a BOTUI that cannot switch to the Tello camera via menu.   If on, then only the Tello camera can be read, preventing the use of the USB camera.

* camera.cpp - make sure that the Tello video has the proper resolution

	Original settings had the tello input being set at HIGH_RES.
However, the tello video is currently fixed at TELLO_RES. At a
minimun this causes issues with the BOTUI display.

Signed-off-by: Eugene Myers <cedarhouse@comcast.net>